% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractData}
\alias{extractData}
\title{extractData}
\usage{
extractData(.spec, .data)
}
\arguments{
\item{.spec}{spec}

\item{.data}{data}
}
\value{
value value
}
\description{
Description extract Data
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

