% Generated by roxygen2 (4.0.1): do not edit by hand
\name{genDataMatrix}
\alias{genDataMatrix}
\title{genDataMatrix}
\usage{
genDataMatrix(data, any_wx = TRUE)
}
\arguments{
\item{data}{data}

\item{any_wx}{any_wx}
}
\value{
value
}
\description{
Description genDataMatrix
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

