% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generateWeighting}
\alias{generateWeighting}
\title{generateWeighting}
\usage{
generateWeighting(D)
}
\arguments{
\item{D}{a network matrix}
}
\value{
weighting matrix
}
\description{
generateWeighting
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}
\keyword{internal}

