\name{genModelData}
\alias{genModelData}
\title{genModelData}
\usage{
genModelData(.spec, save = FALSE, path = "")
}
\arguments{
  \item{.spec}{.spec}

  \item{save}{save}

  \item{path}{path}
}
\value{
value
}
\description{
Description genModelData
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

