% Generated by roxygen2 (4.0.1): do not edit by hand
\name{genNetworkData}
\alias{genNetworkData}
\title{genNetworkData}
\usage{
genNetworkData(data)
}
\arguments{
\item{data}{data}
}
\value{
value
}
\description{
Description genNetworkData
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

