% Generated by roxygen2 (4.0.1): do not edit by hand
\name{genPairwiseHobbyData}
\alias{genPairwiseHobbyData}
\title{genPairwiseHobbyData}
\usage{
genPairwiseHobbyData(H)
}
\arguments{
\item{H}{H}
}
\value{
a vector
}
\description{
genPairwiseHobbyData
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}
\keyword{internal}

