% Generated by roxygen2 (4.0.1): do not edit by hand
\name{genPairwiseIndex}
\alias{genPairwiseIndex}
\title{genPairwiseIndex}
\usage{
genPairwiseIndex(n)
}
\arguments{
\item{n}{}
}
\value{
data.frame
}
\description{
genPairwiseIndex
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}
\keyword{internal}

