\name{genSpecStandardMath}
\alias{genSpecStandardMath}
\title{genSpecStandardMath}
\usage{
genSpecStandardMath(network = c("all", "friends", "studymates"))
}
\arguments{
  \item{network}{network = "all" gives friends and
  studymates networks. network = "friends" gives friends
  network only. network = studymates gives studymates
  network only.}
}
\value{
value
}
\description{
Description genSpecStandardMath
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

