\name{getDataWide}
\alias{getDataWide}
\title{getDataWide}
\usage{
  getDataWide(.raw_data, .school, .drop_by_case_id)
}
\arguments{
  \item{.raw_data}{}

  \item{.school}{}

  \item{.drop_by_case_id}{}
}
\value{
  value
}
\description{
  Description getDataWide
}
\examples{
\dontrun{

}
}
\author{
  TszKin Julian Chan \email{ctszkin@gmail.com}
}

