% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getFileName}
\alias{getFileName}
\title{getFileName}
\usage{
getFileName(type = c("network"), spec)
}
\arguments{
\item{type}{type}

\item{spec}{spec}
}
\value{
value
}
\description{
Description getFileName
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

