\name{getHobby}
\alias{getHobby}
\title{getHobby}
\usage{
  getHobby(.raw_data, .school, .hobby, .drop_by_case_id)
}
\arguments{
  \item{.raw_data}{}

  \item{.school}{}

  \item{.hobby}{}

  \item{.drop_by_case_id}{}
}
\value{
  value
}
\description{
  Description getHobby
}
\examples{
\dontrun{

}
}
\author{
  TszKin Julian Chan \email{ctszkin@gmail.com}
}

