\name{getNetwork}
\alias{getNetwork}
\title{getNetwork}
\usage{
  getNetwork(.raw_data, .school, .network,
    .drop_by_case_id)
}
\arguments{
  \item{.raw_data}{}

  \item{.school}{}

  \item{.network}{}

  \item{.drop_by_case_id}{}
}
\value{
  value
}
\description{
  Description getNetwork
}
\examples{
\dontrun{

}
}
\author{
  TszKin Julian Chan \email{ctszkin@gmail.com}
}

