\name{getNetworkMatrixOrder}
\alias{getNetworkMatrixOrder}
\title{getNetworkMatrixOrder}
\usage{
  getNetworkMatrixOrder(.raw_data, .school)
}
\arguments{
  \item{.raw_data}{}

  \item{.school}{}
}
\value{
  value
}
\description{
  Description getNetworkMatrixOrder
}
\examples{
\dontrun{

}
}
\author{
  TszKin Julian Chan \email{ctszkin@gmail.com}
}

