\name{getPairwiseFriendshipData}
\alias{getPairwiseFriendshipData}
\title{getPairwiseFriendshipData}
\usage{
  getPairwiseFriendshipData(network_data,
    network_formation_formula)
}
\arguments{
  \item{network_data}{network_data}

  \item{network_formation_formula}{network_formation_formula}
}
\value{
  A data.frame consist of the estimates.
}
\description{
  getPairwiseFriendshipData
}
\author{
  TszKin Julian Chan \email{ctszkin@gmail.com}
}
\keyword{internal}

