% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getVariableName}
\alias{getVariableName}
\title{getVariableName}
\usage{
getVariableName(type = c("school", "network"), spec)
}
\arguments{
\item{type}{type}

\item{spec}{spec}
}
\value{
value value
}
\description{
Description getVariableName
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

