\name{prepareData}
\alias{prepareData}
\title{prepareData}
\usage{
prepareData(.raw_data, .spec, .school)
}
\arguments{
  \item{.raw_data}{.raw_data}

  \item{.spec}{.spec}

  \item{.school}{.school}
}
\value{
value
}
\description{
Description prepareDataprepareData
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

