\name{readSpecification}
\alias{readSpecification}
\title{readSpecification}
\usage{
readSpecification(path, type = c("all", "network", "hobby"))
}
\arguments{
  \item{path}{path}

  \item{type}{type}
}
\value{
value
}
\description{
Description readSpecification
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

