\name{splitNetworkData}
\alias{splitNetworkData}
\title{splitNetworkData}
\usage{
  splitNetworkData(network_data, group,
    network_only = FALSE)
}
\arguments{
  \item{network_data}{}

  \item{group}{}

  \item{network_only}{}
}
\value{
  value
}
\description{
  Description splitNetworkData
}
\examples{
\dontrun{

}
}
\author{
  TszKin Julian Chan \email{ctszkin@gmail.com}
}

