% Generated by roxygen2 (4.0.1): do not edit by hand
\name{genModelData}
\alias{genModelData}
\title{genModelData}
\usage{
genModelData(.spec, save = FALSE, path = "", counterfactual_mod)
}
\arguments{
\item{.spec}{.spec}

\item{save}{save}

\item{path}{path}

\item{counterfactual_mod}{counterfactual_mod}
}
\value{
value
}
\description{
Description genModelData
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

