\name{genSpecStandardMath}
\alias{genSpecStandardMath}
\title{genSpecStandardMath}
\usage{
genSpecStandardMath(network = c("all", "friends", "studymates"),
  network_formation_fixed_effect = FALSE)
}
\arguments{
  \item{network}{network = "all" gives friends and
  studymates networks. network = "friends" gives friends
  network only. network = studymates gives studymates
  network only.}

  \item{network_formation_fixed_effect}{Default=FALSE}
}
\value{
value
}
\description{
Description genSpecStandardMath
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

