\name{getFileName}
\alias{getFileName}
\title{getFileName}
\usage{
  getFileName(type = c("network", "hobby"), spec)
}
\arguments{
  \item{type}{}

  \item{spec}{}
}
\value{
  value
}
\description{
  Description getFileName
}
\examples{
\dontrun{

}
}
\author{
  TszKin Julian Chan \email{ctszkin@gmail.com}
}

