\name{getHobby}
\alias{getHobby}
\title{getHobby}
\usage{
getHobby(.raw_data, .school, .hobby, .drop_by_case_id)
}
\arguments{
  \item{.raw_data}{raw_data}

  \item{.school}{school}

  \item{.hobby}{hobby}

  \item{.drop_by_case_id}{drop_by_case_id}
}
\value{
value
}
\description{
Description getHobby
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

