\name{getNetwork}
\alias{getNetwork}
\title{getNetwork}
\usage{
getNetwork(.raw_data, .school, .network, .drop_by_case_id)
}
\arguments{
  \item{.raw_data}{raw_data}

  \item{.school}{school}

  \item{.network}{network}

  \item{.drop_by_case_id}{drop_by_case_id}
}
\value{
value
}
\description{
Description getNetwork
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

