\name{getNetworkMatrixOrder}
\alias{getNetworkMatrixOrder}
\title{getNetworkMatrixOrder}
\usage{
  getNetworkMatrixOrder(.raw_data, .school)
}
\arguments{
  \item{.raw_data}{raw_data}

  \item{.school}{school}
}
\value{
  value
}
\description{
  Description getNetworkMatrixOrder
}
\author{
  TszKin Julian Chan \email{ctszkin@gmail.com}
}

