\name{getVariableName}
\alias{getVariableName}
\title{getVariableName}
\usage{
  getVariableName(type = c("school", "network", "hobby"),
    spec)
}
\arguments{
  \item{type}{}

  \item{spec}{}
}
\value{
  value
}
\description{
  Description getVariableName
}
\examples{
\dontrun{

}
}
\author{
  TszKin Julian Chan \email{ctszkin@gmail.com}
}

