\name{getVariableName}
\alias{getVariableName}
\title{getVariableName}
\usage{
  getVariableName(type = c("school", "network", "hobby"),
    spec)
}
\arguments{
  \item{type}{type}

  \item{spec}{spec}
}
\value{
  value value
}
\description{
  Description getVariableName
}
\author{
  TszKin Julian Chan \email{ctszkin@gmail.com}
}

