% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prepareData}
\alias{prepareData}
\title{prepareData}
\usage{
prepareData(.raw_data, .spec, .school, .counterfactual_mod)
}
\arguments{
\item{.raw_data}{.raw_data}

\item{.spec}{.spec}

\item{.school}{.school}

\item{.counterfactual_mod}{.counterfactual_mod}
}
\value{
value
}
\description{
Description prepareData
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

