\name{readSeshkNetwork}
\alias{readSeshkNetwork}
\title{readSeshkNetwork}
\usage{
readSeshkNetwork(.data_source, .version = "SESHK2011 - network - 0.6.2/")
}
\arguments{
  \item{.data_source}{path}

  \item{.version}{Version number. Default is "SESHK2011 -
  network - 0.6.2/"}
}
\value{
a raw_data
}
\description{
Description
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

