% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readSpecification}
\alias{readSpecification}
\title{readSpecification}
\usage{
readSpecification(path)
}
\arguments{
\item{path}{path}
}
\value{
value
}
\description{
Description readSpecification
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

