\name{readSpecification}
\alias{readSpecification}
\title{readSpecification}
\usage{
  readSpecification(path,
    type = c("all", "network", "hobby"))
}
\arguments{
  \item{path}{}

  \item{type}{}
}
\value{
  value
}
\description{
  Description readSpecification
}
\examples{
\dontrun{

}
}
\author{
  TszKin Julian Chan \email{ctszkin@gmail.com}
}

