\name{splitNetworkData}
\alias{splitNetworkData}
\title{splitNetworkData}
\usage{
  splitNetworkData(network_data, group,
    network_only = FALSE)
}
\arguments{
  \item{network_data}{network_data}

  \item{group}{group}

  \item{network_only}{network_only}
}
\value{
  value
}
\description{
  Description splitNetworkData
}
\author{
  TszKin Julian Chan \email{ctszkin@gmail.com}
}

