% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ToUndirectedGraph}
\alias{ToUndirectedGraph}
\title{ToUndirectedGraph}
\usage{
ToUndirectedGraph(method = c("undirected_and", "undirected_or", "directed",
  "directed_inversed"), network_matrix1, network_matrix2)
}
\arguments{
\item{method}{method}

\item{network_matrix1}{network_matrix1}

\item{network_matrix2}{network_matrix2}
}
\value{
value
}
\description{
Description ToUndirectedGraph
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

